+++ R/query-engine.R
@@ -43,7 +43,7 @@
     ## By default we return also the filter columns!
     columns_tbl <- .reduce_tables_start_from(tbls, columns, start_from)
     paste0(.select(unlist(.prefix_columns(columns_tbl), use.names = FALSE)),
-           .from(names(columns_tbl)),
+           .from(names(columns_tbl), joins = .joins(x)),
            .where(filter, columns_tbl), .order(order))
 }
 
@@ -74,50 +74,84 @@
 #' @md
 #'
 #' @noRd
-.from <- function(tables) {
-    q <- .join_tables(tables)
+.from <- function(tables, joins = .JOINS) {
+    q <- .join_tables(tables, joins = joins)
     paste0(" from ", q)
 }
 
-#' @description Joins two tables from the database.
+.joins <- function(x = new("CompDb")) {
+    if (length(x@.properties[["joins"]]))
+        x@.properties[["joins"]]
+    else .JOINS
+}
+
+#' Initialize that in zzz.R and allow adding joins.
+#'
+#' @noRd
+.JOINS <- rbind(
+    c("ms_compound", "synonym",
+      "on (ms_compound.compound_id=synonym.compound_id)",
+      "left outer join"),
+    c("ms_ion", "ms_compound",
+      "on (ms_ion.compound_id=ms_compound.compound_id)",
+      "left outer join"),
+    c("ms_compound", "msms_spectrum",
+      "on (ms_compound.compound_id=msms_spectrum.compound_id)",
+      "left outer join"),
+    c("ms_ion", "msms_spectrum",
+      "on (ms_ion.compound_id=msms_spectrum.compound_id)",
+      "left outer join"),
+    c("msms_spectrum", "synonym",
+      "on (msms_spectrum.compound_id=synonym.compound_id)",
+      "left outer join"),
+    c("msms_spectrum", "msms_spectrum_peak",
+      "on (msms_spectrum.spectrum_id=msms_spectrum_peak.spectrum_id)",
+      "left outer join")
+)
+
+#' Function to convert the "join definition" into a node list
+#'
+#' @noRd
+.table_to_graph <- function(x) {
+    u <- unique(as.vector(x))
+    res <- lapply(u, function(z) {
+        v <- as.vector(x)[x[, 1L] %in% z | x[, 2L] %in% z]
+        unique(v[!v %in% z])
+    })
+    names(res) <- u
+    res
+}
+
+#' @description
+#'
+#' Creates a join statement to join the database tables which names are
+#' provided with `x`. The function uses `.add_join_tables()` to eventually
+#' add additional database tables in case the ones defined in `x` can not
+#' be directly joined.
 #'
 #' @param x `character` with the names of the tables to be joined.
+#'
+#' @param joins `character` `matrix` with the definition of the joins that
+#'     can be done between database tables.
 #'
 #' @author Johannes Rainer
 #'
 #' @md
 #'
 #' @noRd
-.join_tables <- function(x){
-    .JOINS <- rbind(
-        c("ms_compound", "synonym",
-          "on (ms_compound.compound_id=synonym.compound_id)",
-          "left outer join"),
-        c("ms_ion", "ms_compound",
-          "on (ms_ion.compound_id=ms_compound.compound_id)",
-          "left outer join"),
-        c("ms_compound", "msms_spectrum",
-          "on (ms_compound.compound_id=msms_spectrum.compound_id)",
-          "left outer join"),
-        c("ms_ion", "msms_spectrum",
-          "on (ms_ion.compound_id=msms_spectrum.compound_id)",
-          "left outer join"),
-        c("msms_spectrum", "synonym",
-          "on (msms_spectrum.compound_id=synonym.compound_id)",
-          "left outer join"),
-        c("msms_spectrum", "msms_spectrum_peak",
-          "on (msms_spectrum.spectrum_id=msms_spectrum_peak.spectrum_id)",
-          "left outer join")
-    )
-    x <- .add_join_tables(x)
+.join_tables <- function(x, joins = .JOINS){
+    x <- unique(x)
+    if (length(x) < 2L)
+        return(x)
+    x <- .add_join_tables(x, joins = joins)
     q <- x[1]
     tbls_used <- x[1]
     tbls <- x[-1]
     idxs <- c(1, 2)
     while(length(tbls)) {
-        got_it <- which((.JOINS[, 1] %in% tbls_used & .JOINS[, 2] %in% tbls) |
-                        (.JOINS[, 2] %in% tbls_used & .JOINS[, 1] %in% tbls))
-        join <- .JOINS[got_it[1], ]
+        got_it <- which((joins[, 1] %in% tbls_used & joins[, 2] %in% tbls) |
+                        (joins[, 2] %in% tbls_used & joins[, 1] %in% tbls))
+        join <- joins[got_it[1], ]
         if (length(got_it)) {
             new_tbl <- join[idxs][!(join[idxs] %in% tbls_used)]
             q <- paste(q, join[4], new_tbl, join[3])
@@ -130,13 +164,16 @@
     q
 }
 
-#' @description Helper function to add additional tables required to join the
-#'     provided tables.
+#' @description
 #'
-#' @note This function uses some hard-coded logic based on the database layout
-#'     to define if, and which, tables are needed for a join.
+#' Helper function to add additional tables required to join the two provided
+#' tables.
 #'
-#' @param x `character` with the names of the tables to be joined.
+#' @param x `character` with the names of the tables to be joined. Has to be
+#'     of length > 1.
+#'
+#' @param joins `character` `matrix` with the definition of the tables that
+#'     can be joined.
 #'
 #' @return `character` with all tables needed to join the tables in `x`
 #'     (contain `x` plus eventually required additional tables).
@@ -145,12 +182,20 @@
 #'
 #' @author Johannes Rainer
 #'
+#' @importFrom utils combn
+#'
 #' @noRd
-.add_join_tables <- function(x) {
-    ## msms_spectrum_peak with any other table: need also msms_spectrum_metadata
-    unique(x)
+.add_join_tables <- function(x, joins = .JOINS) {
+    g <- .table_to_graph(joins[, 1:2, drop = FALSE])
+    cmb <- combn(x, 2, simplify = FALSE)
+    res <- unlist(lapply(
+        cmb, function(z) .shortest_path(g, start = z[1L], end = z[2L])),
+        use.names = FALSE)
+    if (!length(res))
+        stop("Unable to join tables ", paste0("'", x, "'", collapse = ", "),
+             call. = FALSE)
+    unique(res)
 }
-
 
 #' @description
 #'
@@ -315,7 +360,7 @@
 
 .prefix_columns <- function(x) {
     mapply(names(x), x, FUN = function(y, z) paste0(y, ".", z),
-                  SIMPLIFY = FALSE)
+           SIMPLIFY = FALSE)
 }
 
 #' Main interface function to retrieve data from the database. Performs the
@@ -354,4 +399,43 @@
             x$intensity <- lapply(x$intensity, unserialize)
     }
     x
+}
+
+#' Counting the number of nodes in a path
+#'
+#' @noRd
+.path_nodes <- function(x = NULL) {
+    if (is.null(x)) return(Inf)
+    length(x)
+}
+
+#' Get the shortest path between two nodes in a graph.
+#'
+#' @param graph `list` of connected nodes. Names of the `list` are node names
+#'     and the elements are the names of the nodes the current node is directly
+#'     connected with.
+#'
+#' @param start `character(1)` with the name of the start node.
+#'
+#' @param end `character(1)` with the name of the end node.
+#'
+#' @param path `character` with the path already travelled. Should be empty
+#'     for the first call. This will be used in the recursive call to keep
+#'     track of the path.
+#'
+#' @noRd
+.shortest_path <- function(graph, start, end, path = c()) {
+    if (is.null(graph[[start]])) return(NULL)
+    path <- c(path, start)
+    ## end of recursion - return path ended in start node
+    if (start == end) return(path)
+    shortest <- NULL
+    for (node in graph[[start]]) {
+        if (!node %in% path) {
+            newpath <- .shortest_path(graph, node, end, path)
+            if (.path_nodes(newpath) < .path_nodes(shortest))
+                shortest <- newpath
+        }
+    }
+    shortest
 }